rewriting comb.py with Classes, now much more readable and useful than beforegit mv comb_new.py comb.py
